Implement concurrent.Referable and concurrent.Ref traits.
Concurrently referable object have acquire and release methods
which ensure that an object being used is never disposed in a
multithreaded environment, as long as some thread called acquired
on the object, but has not called release.

To ease the use of referable objects, Ref cells can be used,
which are essentially wrappers for the references to referable
objects. Refs are thread-safe in the sense that they can be
read and written to by multiple threads.
Linearizibility property is slightly delicate.